feat : css 속성 값 초기화 + 팔레트 색상 초기화
