+++ src/utils/fillSVGPath.ts
@@ -1,0 +1,11 @@
+import { getElement } from "./document";
+
+function fillSVGPath(targetElement: string, targetColor: string) {
+  const SVG = getElement(targetElement);
+  const paths = SVG?.querySelectorAll("path");
+  paths?.forEach((path) => {
+    path.setAttribute("fill", targetColor);
+  });
+}
+
+export default fillSVGPath;
